% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_star_statistics.R
\name{calculate_star_statistics}
\alias{calculate_star_statistics}
\title{Calculates the statistics for the crosssections of a Star}
\usage{
calculate_star_statistics(StarCrosssections, SubsetToUrban = F,
  UrbanLayerName, NDVILayerName, SlopeLayerName, AspectLayerName)
}
\arguments{
\item{StarCrosssections}{a list of crosssections, as created by \code{create_star_crosssections}}

\item{SubsetToUrban}{a switch indicating if the statistics should only be calculated for urban segments. Requires \code{UrbanLayerName} to be defined.}

\item{UrbanLayerName}{name of the column which holds the urban density values}

\item{NDVILayerName}{name of the column which holds the NDVI values}

\item{SlopeLayerName}{name of the column which holds the slope values}

\item{AspectLayerName}{name of the column which holds the aspect values}
}
\value{
Dataframe
}
\description{
Calculates the statistics for the crosssections of a Star
}
\details{
The function calculates descriptive statistics for the distribution of the raster values extracted under a star.

If \code{SubsetToUrban = T} the statistics will be calculated only for the urban segments if the input \code{StarCrosssections} includes a logical column "IsUrban", as created by \code{find_limit_star}.

An \code{UrbanLayerName} is required and returns statistics for the building density distribution.
Optionally, passing \code{NDVILayerName}, \code{SlopeLayerName} and \code{AspectLayerName} will return additional statistics.
All Statistics are calculated for each of the crosssections individually and then averaged over all crosssections, weighted by their length. The exception is MeanUrbanCentralOffset, for which all crosssections are equally weighted.


  Calculated statistics include:
 \enumerate{
  \item MeanUrbanMean, the weighted mean of the means of the building density
  \item MeanUrbanSkew, the weighted mean of the skews of the building density
  \item MeanUrbanKurtosis, the weighted mean of the kurtosis of the building density
  \item MeanUrbanVariance, the weighted mean of the variances of the building density
  \item MeanUrbanCentralOff, the mean number of segments by which the maximum of the curves differ from the centers of the building density distributions.
  \item MeanGreenMean, the weighted mean of the means of the NDVI distributions
  \item MeanUrbanGreenness, the weighted mean of the scalar products of the building density and the NDVI distributions.
  \item MeanUrbanGreennessCor, the weighted mean of the correlations of the building density and the NDVI distributions
  \item MeanUrbanSlope, the weighted mean of the scalar products of the building density and the slope distributions.
  \item MeanSlopeMean, the weighted mean of the means of the slope
  \item MeanUrbanAspect, the weighted mean of the means of the aspect
}
}
\examples{
  calculate_star_statistics(
  StarCrosssections = StarCross,
  SubsetToUrban = F,
  UrbanLayerName = "GUFNorm",
  NDVILayerName = "NDVI",
  SlopeLayerName = "slope",
  AspectLayerName = "aspect")

}
\seealso{
\code{\link{create_star_crosssections}},
\code{\link{find_limit}}
}
\author{
Johannes Mast
}
