% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_beam.R
\name{create_beam}
\alias{create_beam}
\title{Create a beam starting from a central point}
\usage{
create_beam(UrbanCenter, BeamWidth = 10000, BeamLength = 70000,
  NSegments = 10, OffAngle = 0)
}
\arguments{
\item{UrbanCenter}{a SpatialPoint object}

\item{BeamWidth}{width of the beam in meters}

\item{BeamLength}{length of the beam in meters}

\item{NSegments}{number of segments of the beam}

\item{OffAngle}{azimuth of the beam}
}
\value{
SpatialPolygonsDataframe
}
\description{
Create a beam starting from a central point
}
\details{
This function creates a beam object starting from a given \code{UrbanCenter}.
The returned Beam is a SpatialPolygonsDataFrame and is unprojected in WGS84 (EPSG:4326).
It consists of \code{NSegments} equally sized polygons of length \code{BeamLength/NSegments} and width \code{BeamWidth}. The \code{UrbanCenter} lies on the midpoint of the innermost edge of the innermost segment. From there, the beam extends for \code{BeamLength} meters at a true-north-azimuth of \code{OffAngle}, the edges of the beam forming a geodesic between its corners.
}
\examples{
create_beam(UrbanCenter)

}
\seealso{
\code{\link{create_beam}},
\code{\link{create_star_click}},
\code{\link[geosphere]{dest_point}}
}
\author{
Johannes Mast
}
