% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_star.R
\name{create_star}
\alias{create_star}
\title{Create a Star object around a central point}
\usage{
create_star(UrbanCenter, BeamWidth = 10000, BeamLength = 70000,
  NSegments = 10, NBeams = 8)
}
\arguments{
\item{UrbanCenter}{a SpatialPoint object}

\item{BeamWidth}{width of each beam in meters}

\item{BeamLength}{length of each beam in meters}

\item{NSegments}{number of segments of each beam}

\item{NBeams}{the number of beams of the star}
}
\value{
SpatialPolygonsDataframe
}
\description{
Create a Star object around a central point
}
\details{
This function creates a Star object around a given \code{UrbanCenter}.
It consists of \code{NBeams} times \code{NSegments} equally sized polygons of length \code{BeamLength/NSegments} and width \code{BeamWidth},
arranged in \code{NBeams} equally spaced beams of length \code{Beamlength}.
 The returned Star is a SpatialPolygonsDataFrame and is unprojected in WGS84 (EPSG:4326).
 It contains a field "OverallId" which contains unique ids for each segment, the first 6 numbers indicating the beam, the latter 6 numbers indicating the segment.
}
\examples{
create_star_click(UrbanCenter)

}
\seealso{
\code{\link{create_beam}},
\code{\link{create_star_click}}
}
\author{
Johannes Mast
}
