% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_star_crosssections.R
\name{create_star_crosssections}
\alias{create_star_crosssections}
\title{Create crosssections from a Star object}
\usage{
create_star_crosssections(Star)
}
\arguments{
\item{Star}{a Star Object that contains columns "BeamId" and "SegmentId", as created by \code{create_star()}}
}
\value{
List of dataframes
}
\description{
Create crosssections from a Star object
}
\details{
Splits the dataframe ofa \code{Star} into its constituent beams (identified by their "BeamId"). Binds the data of each beam to the reverse of its opposite beam. This creates crosssections with the original geographic center of the \code{Star} in the center of the crosssections.


  The number of beams must be an even number.
}
\examples{
create_star_crosssections(Star)

}
\seealso{
\code{\link{create_star}},
\code{\link{split_star_into_beams}}
}
\author{
Johannes Mast
}
