% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_under_star.R
\name{extract_under_star}
\alias{extract_under_star}
\title{Extract raster layers under Star object}
\usage{
extract_under_star(Star, LayerList, funs = mean)
}
\arguments{
\item{Star}{a star object}

\item{LayerList}{a list object of raster layers}

\item{funs}{a function which is passed to raster::extract()}
}
\value{
Dataframe
}
\description{
Extract raster layers under Star object
}
\details{
This Function extracts the values of a \code{LayerList} of raster layers under a \code{Star}
 and returns a data frame in which the rows correspond to the Stars segments and the columns to the extracted raster layers.
 Optionally applies a function \code{funs} during the extraction.
}
\examples{
find_limit_star(Star,UrbanLayerName,Threshold=50,SlopeThreshold=0.01,SlopeWindow=3)

}
\seealso{
\code{\link[raster]{extract}}
}
\author{
Johannes Mast
}
