% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_limit.R
\name{find_limit}
\alias{find_limit}
\title{Find the limit of a unilateral density distribution}
\usage{
find_limit(BeamSeries, Threshold = 0.2, SlopeThreshold = 1e-04,
  SlopeWindow = 3)
}
\arguments{
\item{BeamSeries}{a unilateral density distribution}

\item{Threshold}{a threshold under which the density is considered nonurban}

\item{SlopeThreshold}{the threshold under which the in SlopeWindow calculated mean of the slope is considered to be small}

\item{SlopeWindow}{a forward looking search window in which the mean slope is evaluated}
}
\value{
Integer
}
\description{
Find the limit of a unilateral density distribution
}
\details{
Take a \code{BeamSeries} of density values of a beam and finds a plausible segmentation point, which can be considered the urban limit.
The criterium for the point being: First point at which
\enumerate{
\item the value is lower than \code{"Threshold"} and
\item the mean slope over the next \code{SlopeWindow} data points is lower than \code{SlopeThreshold}
}
If no point satisfies both conditions, NA is returned.
}
\seealso{
\code{\link{find_limit_star}}
}
\author{
Johannes Mast
}
