% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_limit_star.R
\name{find_limit_star}
\alias{find_limit_star}
\title{Find the limits of a Stars density distributions}
\usage{
find_limit_star(Star, UrbanLayerName, Threshold = 50,
  SlopeThreshold = 0.01, SlopeWindow = 3)
}
\arguments{
\item{Star}{a Star object that contains columns "BeamId" and "SegmentId", as created by create_star()}

\item{UrbanLayerName}{the name of the column which holds the density values}

\item{Threshold}{a threshold under which the density is considered nonurban}

\item{SlopeThreshold}{the threshold under which the in SlopeWindow calculated mean of the slope is considered to be small}

\item{SlopeWindow}{a forward looking search window in which the mean slope is evaluated}
}
\value{
Logical vector
}
\description{
Find the limits of a Stars density distributions
}
\details{
Splits the dataframe of a \code{Star} into its constituent beams (identified by their "BeamId").
Each beams \code{UrbanLayerName} column is then evaluated to find a plausible segmentation point, which can be considered the urban limit.
The criterium for the point being: First point at which
\enumerate{
\item the value is lower than \code{"Threshold"} and
\item the mean slope over the next \code{SlopeWindow} data points is lower than \code{SlopeThreshold}
}
The limits are then used to create a logical vector of \code{length(Star)}, indicating for each segment of the Star whether it is urban.
If for at least one of the beams, no point satisfies both conditions, NA is returned.
}
\examples{
find_limit_star(Star,UrbanLayerName,Threshold=50,SlopeThreshold=0.01,SlopeWindow=3)


}
\seealso{
\code{\link{find_limit}},
\code{\link{split_star_into_beams}}
}
\author{
Johannes Mast
}
