% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_star_into_beams.R
\name{split_star_into_beams}
\alias{split_star_into_beams}
\title{Split the dataframe of a Star object into its constituent beams}
\usage{
split_star_into_beams(Star)
}
\arguments{
\item{Star}{a Star object that contains columns "BeamId" and "SegmentId", as created by create_star()}
}
\value{
List of dataframes
}
\description{
Split the dataframe of a Star object into its constituent beams
}
\details{
Splits the dataframe of a \code{Star} into its constituent beams (identified by their "BeamId"). Each resulting beam dataframe is as long as the number of segments of the beam.
}
\examples{
split_star_into_beams(Star)
}
\seealso{
\code{\link{create_star}}
}
\author{
Johannes Mast
}
